Khoa
fix logic nhung dang khong ket noi duoc voi database

+++ Smoking-Cessation-Support-Platform/src/main/java/com/smokingcessation/controller/SubscriptionController.java
@@ -30,7 +30,7 @@ public class SubscriptionController {
     }
 
     @GetMapping("/member/{memberId}")
-    public ResponseEntity<List<Subscription>> getByMemberId(@PathVariable Integer memberId) {
+    public ResponseEntity<List<Subscription>> getByMemberId(@PathVariable Long memberId) {
         return ResponseEntity.ok(subscriptionService.getSubscriptionsByMemberId(memberId));
     }
