+++ Smoking-Cessation-Support-Platform/src/main/java/com/smokingcessation/service/SubscriptionService.java
@@ -28,7 +28,7 @@ public class SubscriptionService {
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy subscription"));
     }
 
-    public List<Subscription> getSubscriptionsByMemberId(Integer memberId) {
+    public List<Subscription> getSubscriptionsByMemberId(Long memberId) {
         return subscriptionRepository.findByMember_MemberID(memberId);
     }
